-Implementazione funzione che ritorna le auto preotabili in un determinato periodo. -Implementazione approvazione/declinazione prenotazioni. -Implementazione modifica/elimina prenotazione se la data di inizio dista almeno 2 giorni da quella odierna.

+++ src/main/java/com/stage/rentalcar/services/ReservationServiceImpl.java
@@ -10,8 +10,12 @@ import com.stage.rentalcar.repository.ReservationRepository;
 import com.stage.rentalcar.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
+import java.time.temporal.ChronoUnit;
 import java.util.List;
 
 @Service
@@ -22,7 +26,8 @@ public class ReservationServiceImpl implements ReservationService {
     private final ReservationMapper reservationMapper;
     private final UserRepository userRepository;
     private final CarRepository carRepository;
-
+    private final CarService carService;
+    private static final Logger logger = LoggerFactory.getLogger(ReservationServiceImpl.class);
 
     @Override
     public List<Reservation> getReservationsForUser(User user) {
@@ -35,6 +40,22 @@ public class ReservationServiceImpl implements ReservationService {
     }
 
     @Override
+    public List<Reservation> getReservationsBetweenDates(LocalDate start, LocalDate end) {
+        return reservationRepository.getReservationsBetweenDates(start, end);
+    }
+
+    @Override
+    public List<Car> getFreeCars(ReservationDTO reservationDTO) throws Exception {
+        long days = ChronoUnit.DAYS.between(LocalDate.now(), reservationDTO.getStartDate());
+        if(days>2){
+            List<Reservation> resBetweenDates = getReservationsBetweenDates(reservationDTO.getStartDate(), reservationDTO.getEndDate());
+            return carService.getFreeCars(resBetweenDates);
+        } else {
+            throw new Exception("Mancano meno di 2 giorni alla data di inizio.");
+        }
+    }
+
+    @Override
     public void insOrUpReservation(ReservationDTO reservationDTO) {
         User user = userRepository.getById(reservationDTO.getUserId());
         Car car = carRepository.getById(reservationDTO.getCarId());
@@ -42,7 +63,31 @@ public class ReservationServiceImpl implements ReservationService {
     }
 
     @Override
-    public void delReservation(Integer id) {
-        reservationRepository.deleteById(id);
+    public void approveReservation(Integer id) {
+        ReservationDTO reservationDTO = getReservationById(id);
+        Reservation reservation = reservationMapper.fromDTOtoEntity(reservationDTO, userRepository.getById(reservationDTO.getUserId()),
+                                carRepository.getById(reservationDTO.getCarId()));
+        reservation.setConfirmed(true);
+        reservationRepository.saveAndFlush(reservation);
+    }
+
+    @Override
+    public void declineReservation(Integer id) {
+        ReservationDTO reservationDTO = getReservationById(id);
+        Reservation reservation = reservationMapper.fromDTOtoEntity(reservationDTO, userRepository.getById(reservationDTO.getUserId()),
+                carRepository.getById(reservationDTO.getCarId()));
+        reservation.setConfirmed(false);
+        reservationRepository.saveAndFlush(reservation);
+    }
+
+    @Override
+    public void delReservation(Integer id) throws Exception {
+        ReservationDTO reservationDTO = getReservationById(id);
+        long days = ChronoUnit.DAYS.between(LocalDate.now(), reservationDTO.getStartDate());
+        if(days>2){
+            reservationRepository.deleteById(id);
+        } else {
+            throw new Exception("Mancano meno di due giorni alla data di inizio.");
+        }
     }
 }
